+++ src/Vortice.Dxc/IDxcCompilerArgs.cs
@@ -5,57 +5,82 @@ namespace Vortice.Dxc;
 
 public unsafe partial class IDxcCompilerArgs
 {
-    public Result AddArguments(string[] arguments)
+    private string[]? _arguments;
+
+    public char** GetArgumentsPtr()
     {
-        return AddArguments(arguments, arguments.Length);
+        const int GetArguments__vtbl_index = 3;
+        return ((delegate* unmanaged[MemberFunction]<IntPtr, char**>)this[GetArguments__vtbl_index])(NativePointer);
+    }
+
+    public string[] Arguments
+    {
+        get
+        {
+            if (_arguments == null || _arguments.Length != Count)
+            {
+                _arguments = new string[Count];
+
+                char** argsPtr = GetArgumentsPtr();
+                for (int i = 0; i < _arguments.Length; i++)
+                {
+                    _arguments[i] = new string(argsPtr[i]);
+                }
+            }
+
+            return _arguments!;
+        }
+    }
+
+    public Result AddArguments(params string[] arguments)
+    {
+        Utf16PinnedStringArray native = new(arguments);
+        try
+        {
+            return AddArguments(native.Handle, native.Length);
+        }
+        finally
+        {
+            native.Release();
+        }
     }
 
     public Result AddArguments(string[] arguments, int argumentsCount)
     {
-        IntPtr* argumentsPtr = (IntPtr*)0;
-
+        Utf16PinnedStringArray native = new(arguments, argumentsCount);
         try
         {
-            if (arguments != null && argumentsCount > 0)
-            {
-                argumentsPtr = Interop.AllocToPointers(arguments, argumentsCount);
-            }
-
-            return AddArguments((IntPtr)argumentsPtr, argumentsCount);
+            return AddArguments(native.Handle, native.Length);
         }
         finally
         {
-            if (argumentsPtr != null)
-            {
-                NativeMemory.Free(argumentsPtr);
-            }
+            native.Release();
         }
     }
 
-    public Result AddArgumentsUTF8(string[] arguments)
+    public Result AddArgumentsUTF8(params string[] arguments)
     {
-        return AddArgumentsUTF8(arguments, arguments.Length);
-    }
-
-    public unsafe Result AddArgumentsUTF8(string[] arguments, int argumentsCount)
-    {
-        IntPtr* argumentsPtr = (IntPtr*)0;
-
+        Utf8PinnedStringArray native = new(arguments);
         try
         {
-            if (arguments != null && argumentsCount > 0)
-            {
-                argumentsPtr = Interop.AllocToPointers(arguments, argumentsCount);
-            }
-
-            return AddArgumentsUTF8((IntPtr)argumentsPtr, argumentsCount);
+            return AddArgumentsUTF8(native.Handle, native.Length);
         }
         finally
         {
-            if (argumentsPtr != null)
-            {
-                NativeMemory.Free(argumentsPtr);
-            }
+            native.Release();
+        }
+    }
+
+    public Result AddArgumentsUTF8(string[] arguments, int argumentsCount)
+    {
+        Utf8PinnedStringArray native = new(arguments, argumentsCount);
+        try
+        {
+            return AddArgumentsUTF8(native.Handle, native.Length);
+        }
+        finally
+        {
+            native.Release();
         }
     }
 }